+++ CHANGES.rst
@@ -21,6 +21,10 @@ The ASDF Standard is at v1.6.0
   is removed in an upcoming asdf release will be ``False`` and
   asdf will no longer by-default memory map arrays. [#1667]
 
+- Introduce ``asdf.util.load_yaml`` to load just the YAML contents
+  of an ASDF file (with the option ``tagged`` to load the contents
+  as a tree of ``asdf.tagged.Tagged`` instances to preserve tags) [#1700]
+
 3.0.1 (2023-10-30)
 ------------------